create utils.py with helper function
Create a utils.py file that defines the SSL helper function that is called in demographics.py and in macro_params.py. Changed the code in these two scripts to import the utils.py file and calls the new function when accessing data from the UN Data Portal.
