+++ ogzaf/demographics.py
@@ -28,8 +28,7 @@ from ogcore import parameter_plots as pp
 from pandas_datareader import wb
 import matplotlib.pyplot as plt
 import requests
-import urllib3
-import ssl
+from ogzaf import utils
 from io import StringIO
 
 # create output director for figures
@@ -38,41 +37,6 @@ DATA_DIR = os.path.join(CUR_PATH, "data", "demographic")
 OUTPUT_DIR = os.path.join(CUR_PATH, "OUTPUT", "Demographics")
 if os.access(OUTPUT_DIR, os.F_OK) is False:
     os.makedirs(OUTPUT_DIR)
-
-"""
-------------------------------------------------------------------------
-Define functions
-------------------------------------------------------------------------
-"""
-
-"""
-The UN Data Portal server doesn't support "RFC 5746 secure renegotiation". This causes and error when the client is using OpenSSL 3, which enforces that standard by default.
-The fix is to create a custom SSL context that allows for legacy connections. This defines a function get_legacy_session() that should be used instead of requests().
-"""
-
-
-class CustomHttpAdapter(requests.adapters.HTTPAdapter):
-    # "Transport adapter" that allows us to use custom ssl_context.
-
-    def __init__(self, ssl_context=None, **kwargs):
-        self.ssl_context = ssl_context
-        super().__init__(**kwargs)
-
-    def init_poolmanager(self, connections, maxsize, block=False):
-        self.poolmanager = urllib3.poolmanager.PoolManager(
-            num_pools=connections,
-            maxsize=maxsize,
-            block=block,
-            ssl_context=self.ssl_context,
-        )
-
-
-def get_legacy_session():
-    ctx = ssl.create_default_context(ssl.Purpose.SERVER_AUTH)
-    ctx.options |= 0x4  # OP_LEGACY_SERVER_CONNECT  #in Python 3.12 you will be able to switch from 0x4 to ssl.OP_LEGACY_SERVER_CONNECT.
-    session = requests.session()
-    session.mount("https://", CustomHttpAdapter(ctx))
-    return session
 
 
 def get_un_fert_data(
